Move keyring and secret handling into a separate module
